+++ src/main/java/com/project/runningcrew/exception/notFound/MemberNotFoundException.java
@@ -1,4 +1,4 @@
-package com.project.runningcrew.exception;
+package com.project.runningcrew.exception.notFound;
 
 public class MemberNotFoundException extends ResourceNotFoundException{
     public MemberNotFoundException() {
