+++ src/main/java/com/project/runningcrew/exception/notFound/RunningRecordNotFoundException.java
@@ -1,4 +1,4 @@
-package com.project.runningcrew.exception;
+package com.project.runningcrew.exception.notFound;
 
 public class RunningRecordNotFoundException extends ResourceNotFoundException{
     public RunningRecordNotFoundException() {
