+++ src/test/java/com/project/runningcrew/service/CrewServiceTest.java
@@ -6,7 +6,7 @@ import com.project.runningcrew.entity.areas.GuArea;
 import com.project.runningcrew.entity.members.Member;
 import com.project.runningcrew.entity.members.MemberRole;
 import com.project.runningcrew.entity.users.User;
-import com.project.runningcrew.exception.CrewNotFoundException;
+import com.project.runningcrew.exception.notFound.CrewNotFoundException;
 import com.project.runningcrew.repository.CrewRepository;
 import com.project.runningcrew.repository.MemberRepository;
 import com.project.runningcrew.service.images.ImageService;
